+++ include/rawdisk.h
@@ -38,8 +38,8 @@ int free_buffer(void* buffer);
 int open_disk(const char* disk_name);
 int close_disk();
 
-int read_block_raw(uint8_t* buffer, uint32_t block_number);
-int write_block_raw(const uint8_t* buffer, uint32_t block_number);
+int read_block_raw(uint8_t* buffer, uint64_t block_number);
+int write_block_raw(const uint8_t* buffer, uint64_t block_number);
 
 int fsync_disk();
